Add adversarial and baseline CLI
